% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotate Peaks}
\usage{
annotatePeaks(peaks)
}
\arguments{
\item{peaks}{A GRanges object containing DRIP-Seq peaks}
}
\value{
A GRanges object containing annotated DRIP-Seq peaks
}
\description{
Annotates DRIP-Seq peaks as a GRanges object with gene-level information
}
