% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateRLoops.R
\name{annotateRLoops}
\alias{annotateRLoops}
\title{Annotate R-Loops}
\usage{
annotateRLoops(peaks)
}
\arguments{
\item{peaks}{A GRanges object containing DRIP-Seq peaks}
}
\value{
A GRanges object containing annotated DRIP-Seq peaks
}
\description{
Annotates DRIP-Seq peaks as a GRanges object with gene-level information
}
