% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotationLst}
\alias{annotationLst}
\title{List of list of tibbles. Containins genomic annotations for use with RSeqR.
See data-raw/anotationLst.R for details.}
\format{
An object of class \code{list} of length 2.
}
\usage{
annotationLst
}
\description{
List of list of tibbles. Containins genomic annotations for use with RSeqR.
See data-raw/anotationLst.R for details.
}
\keyword{datasets}
