% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aux}
\alias{aux}
\title{Auxilliary data for use with RLSeq
Generated via aux.R script in data-raw/}
\format{
An object of class \code{list} of length 7.
}
\usage{
aux
}
\description{
Auxilliary data for use with RLSeq
Generated via aux.R script in data-raw/
}
\keyword{datasets}
