% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatasets.R
\name{buildAvailableGenomes}
\alias{buildAvailableGenomes}
\title{Build Available Genomes
Helper Function which builds the UCSC available genomes dataset}
\usage{
buildAvailableGenomes(...)
}
\arguments{
\item{...}{arguments passed to \code{getEffectiveGenomeSizes()}}
}
\value{
A data frame of avaialable genomes and corresponding info.
}
\description{
Build Available Genomes
Helper Function which builds the UCSC available genomes dataset
}
