% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildCondaEnv}
\alias{buildCondaEnv}
\title{builds the conda env through reticulate}
\usage{
buildCondaEnv(envName, ...)
}
\arguments{
\item{envName}{name for new conda env}

\item{...}{other arguments passed to reticulate::conda_create()}
}
\value{
File path to created environment
}
\description{
builds the conda env through reticulate
}
