% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildGenomeMasks}
\alias{buildGenomeMasks}
\title{Build Genome Masks
Helper Function which builds GRanges of masked genomes for RegioneR}
\usage{
buildGenomeMasks()
}
\value{
A named list of GRanges object with the masked chrom sizes.
}
\description{
Build Genome Masks
Helper Function which builds GRanges of masked genomes for RegioneR
}
