% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkGenes}
\alias{checkGenes}
\title{Check Gene Annotations
Helper function which checks the chrom sizes info from UCSC}
\usage{
checkGenes(genome)
}
\arguments{
\item{genome}{the UCSC genome for which to download chrom sizes}
}
\value{
logical. TRUE if status code 200, FALSE if not
}
\description{
Check Gene Annotations
Helper function which checks the chrom sizes info from UCSC
}
