% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkHomer}
\alias{checkHomer}
\title{Check Homer Annotations
Helper function which checks whether homer annotations are available}
\usage{
checkHomer(genome)
}
\arguments{
\item{genome}{the UCSC genome for which to download chrom sizes}
}
\value{
logical. TRUE if status code 200, FALSE if not
}
\description{
Check Homer Annotations
Helper function which checks whether homer annotations are available
}
