% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRLFSAnno}
\alias{checkRLFSAnno}
\title{Check RLFS}
\usage{
checkRLFSAnno(genome)
}
\arguments{
\item{genome}{the UCSC genome name to check}
}
\value{
A logical, TRUE if available, FALSE if not
}
\description{
Helper function that checks whether a genome has RLFS available.
}
