% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature_ggplot}
\alias{feature_ggplot}
\title{Feature ggplot}
\usage{
feature_ggplot(x, usamp, limits, splitby)
}
\arguments{
\item{x}{A \code{tbl} containing data for plotting.}

\item{usamp}{The name of the user-supplied sample}

\item{limits}{Specify limits on data. Useful for controlling infinite
estimation of odds ratio
resulting from fisher's exact test. To remove limits, set c(-Inf, Inf).
Default: c(-10, 15).}

\item{splitby}{Metadata by which to split plots. Can be "none", "prediction",
or "label".}
}
\value{
A ggplot2 object.
}
\description{
The core plotting component of \code{plotEnrichment}
}
