% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fftModel}
\alias{fftModel}
\title{A Caret model used for predicting whether a peakset resembles
"case" (normal) or "control" (RNaseH1-like) conditions.
Used in predictionCondition().}
\format{
An object of class \code{caretStack} of length 3.
}
\usage{
fftModel
}
\description{
A Caret model used for predicting whether a peakset resembles
"case" (normal) or "control" (RNaseH1-like) conditions.
Used in predictionCondition().
}
\keyword{datasets}
