% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Annotate R-Loops with Genes}
\usage{
geneAnnotation(peaks)
}
\arguments{
\item{peaks}{A GRanges object containing R-loop peaks}
}
\value{
A GRanges object containing annotated R-loop peaks
}
\description{
Annotates R-loop peaks as a GRanges object with gene-level information
}
\examples{

geneAnnotation(RSeqR::SRX1025890_peaks)

}
