% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genomeMasks}
\alias{genomeMasks}
\title{List of GRanges. Contains genomic masks for several commons species.
Generated via genome_masks.R script in data-raw/}
\format{
An object of class \code{list} of length 8.
}
\usage{
genomeMasks
}
\description{
List of GRanges. Contains genomic masks for several commons species.
Generated via genome_masks.R script in data-raw/
}
\keyword{datasets}
