% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChain}
\alias{getChain}
\title{Get Chain
Helper function that retrieves chain file for liftUtil()}
\usage{
getChain(genomeFrom, genomeTo)
}
\arguments{
\item{genomeFrom}{the UCSC genome name to convert from.}

\item{genomeTo}{the UCSC genome name to convert to.}
}
\description{
Get Chain
Helper function that retrieves chain file for liftUtil()
}
