% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSizes}
\alias{getChromSizes}
\title{Get Chrom Sizes
Helper function which downloads chrom sizes from UCSC for a genome.}
\usage{
getChromSizes(genome)
}
\arguments{
\item{genome}{the UCSC genome for which to download chrom sizes}
}
\value{
A tibble containing chrom sizes
}
\description{
Get Chrom Sizes
Helper function which downloads chrom sizes from UCSC for a genome.
}
