% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEffectiveGenomeSizes}
\alias{getEffectiveGenomeSizes}
\title{Get list of effective genome sizes}
\usage{
getEffectiveGenomeSizes(
  genome,
  envPath,
  lengths = c(36, 50, 75, 100, 125, 150, 200, 250, 300)
)
}
\arguments{
\item{genome}{the UCSC genome name to check}

\item{envPath}{the path to conda env containing "khmer"}
}
\value{
A data frame containing the effective genome sizes.
}
\description{
Get list of effective genome sizes
}
