% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEffGenSize}
\alias{getEffGenSize}
\title{Get effective genome size}
\usage{
getEffGenSize(len, faFile, envPath)
}
\arguments{
\item{len}{The read length for which to obtain genome sizes}

\item{faFile}{The genome fasta file to analyze}

\item{envPath}{The path to the conda env with khmer installed}

\item{condaPath}{The path to the conda binary. Default = reticulate::conda_binary()}
}
\value{
A data frame containing the effective genome sizes.
}
\description{
Get effective genome size
}
