% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGSSignal}
\alias{getGSSignal}
\title{Get GS Signal}
\usage{
getGSSignal(coverage, genome)
}
\arguments{
\item{coverage}{The path to a .bigWig file (can be a URL)}

\item{genome}{The UCSC genome ID to use. (Currently, only "hg38" is supported)}
}
\value{
A named list containing the results of correlation analysis.
}
\description{
Extract signal around GS R-loop sites
}
