% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHomerAnno}
\alias{getHomerAnno}
\title{Get Homer Anno
Helper function that retrieves Homer Annotations}
\usage{
getHomerAnno(genome)
}
\arguments{
\item{genome}{the UCSC genome name to retrieve annotations for}
}
\value{
A list object with annotations for that species.
}
\description{
Get Homer Anno
Helper function that retrieves Homer Annotations
}
