% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMask2}
\alias{getMask2}
\title{Build Mask for Genome as GRanges
Adapted from \code{regioneR::getMask()} in the}
\usage{
getMask2(genome)
}
\value{
A GRanges object with the masked chrom sizes.
}
\description{
Build Mask for Genome as GRanges
Adapted from \code{regioneR::getMask()} in the
}
