% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRLFS}
\alias{getRLFS}
\title{Get RLFS}
\usage{
getRLFS(genome)
}
\arguments{
\item{genome}{the UCSC genome for which to download RLFS}
}
\value{
A GRanges object of RLFS.
}
\description{
Helper function which obtains RLFS for a given genome.
}
