% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsSignalRMapDB}
\alias{gsSignalRMapDB}
\title{The matrix generated by summarizing all hg38 RMapDB bigWig files around the
"Gold-Standard" R-loop sites defined by Chedin et al. https://doi.org/10.15252/embj.2020106394
Columns are sample names and rows are hg38 genomic coordinates.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3434 rows and 300 columns.
}
\usage{
gsSignalRMapDB
}
\description{
The matrix generated by summarizing all hg38 RMapDB bigWig files around the
"Gold-Standard" R-loop sites defined by Chedin et al. https://doi.org/10.15252/embj.2020106394
Columns are sample names and rows are hg38 genomic coordinates.
}
\keyword{datasets}
