% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{liftUtil}
\alias{liftUtil}
\title{Lift Over Utility}
\usage{
liftUtil(ranges, genomeFrom, genomeTo)
}
\arguments{
\item{ranges}{A GRanges object in hg19}

\item{genomeFrom}{Genome of ranges supplied, in UCSC format (e.g., "hg19")}

\item{genomeTo}{Genome to convert to (e.g., "hg38")}
}
\value{
A lifted GRanges object
}
\description{
Convenience function for converting ranges from between assemblies
}
\examples{

hg38Lift(RSeqR::SRX1025890_peaks_hg19)

}
