% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Generate report}
\usage{
makeReport(data, outputFile = "RSeqReport.html", ...)
}
\arguments{
\item{data}{A named list containing \code{annotated_peaks}, \code{feature_overlaps},
and \code{rlfs_results}}

\item{outputFile}{A path indicating the report output HTML file.}

\item{...}{Arguments passed to \code{rmarkdown::render()}}
}
\description{
Generates RSeqR Quality Report
}
\examples{

RSeqR::makeReport(RSeqR::SRX1025890, outputFile = "report.html")

}
