% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modes}
\alias{modes}
\title{A tibble containing the R-loop-mapping modes and information about them.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 7 columns.
}
\usage{
modes
}
\description{
A tibble containing the R-loop-mapping modes and information about them.
}
\keyword{datasets}
