% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseAnalyze.R
\name{noiseAnalyze}
\alias{noiseAnalyze}
\title{Analyze sample noise}
\usage{
noiseAnalyze(object, windows = NULL, force = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{windows}{Genomics windows to use for quantifying signal.
Will be automatically supplied if not provided. It is recommended NOT
to specify this option for most analysis types, as doing so will impair
the ability to compare to RLBase samples. Default: NULL.}

\item{force}{Force \code{noiseAnalyze} to run, even if on Windows. Default: FALSE.}
}
\value{
An RLRanges object with noise analysis results included as a \code{tbl}.
The result is accessed via
\code{rlresults(object, "noiseAnalysis")}.
}
\description{
Analyzes the noiseness of the supplied sample using the method described by
\emph{Diaz et al.}. See \emph{details}.
}
\details{
Currently, this does not work on windows.
\subsection{Method}{

The method used for noise analysis is a minor modification of the method
developed by \href{https://pubmed.ncbi.nlm.nih.gov/22499706/}{Diaz et al., 2012}
and also implemented by the
\href{https://deeptools.readthedocs.io/en/develop/}{deepTools} function,
\href{https://deeptools.readthedocs.io/en/develop/content/tools/plotFingerprint.html}{plotFingerprint}.

Briefly, if user-supplied \link{RLRanges} contain a bigWig coverage file,
then the coverage is quantified within random genomic regions
(\code{randomWindows}). The regions are then ranked. A good signal-to-noise
ratio will yield a distribution where most bins have little coverage
but a few have very high coverage. Use downstream tools like
\code{plotNoise} and \code{plotCompareNoise} to visualize these results.
}
}
\examples{

# Example RLRanges object
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# noiseAnalyze does not work on Windows OS
if (.Platform$OS.type != "windows") {
    # run noiseAnalyze
    rlr <- noiseAnalyze(rlr)
}

}
