% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{noiseComparisonPlot}
\alias{noiseComparisonPlot}
\title{Creates a metaplot for comparing noise analysis results with RLBase}
\usage{
noiseComparisonPlot(object, mode = "auto", simple = TRUE, returnData = FALSE)
}
\arguments{
\item{object}{An RLRanges object with \link{noiseAnalyze} already run.}

\item{mode}{A \code{character} containing the R-loop data mode to compare
against. See details for more information.}

\item{simple}{A \code{logical} which specifies whether the plot should only show
samples where the prediction and label are the same. Default: TRUE.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object or a \code{tbl} if \code{returnData} is \code{TRUE}.
}
\description{
Plots the average standardized signal from \link{noiseAnalyze} alongside the
samples in RLBase. For this plot, lower average signal indicates better
signal to noise ratio. \strong{Note}: This plot may be misleading if you supplied
custom windows when running \link{noiseAnalyze}.
}
\details{
\subsection{Mode}{

The \code{mode} parameter specifies the R-loop modality to compare the
user-supplied sample against in the plot. The default, \code{"auto"}
specifies that the \code{mode} from the supplied \code{RLRanges} object will
be used. Only one mode can be specified. For a list of applicable modes,
see \code{auxdata$available_modes}.
}

\subsection{Plot}{

The plot is a violin / jitter plot showing the distribution of average values
from the \link{noiseAnalyze} output across RLBase samples of the selected \code{mode}.
The user-supplied sample is annotated on the plot.
}
}
\examples{

rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Plot RL-Region overlap
noiseComparisonPlot(rlr)

# Return data only
noiseComparisonPlot(rlr, returnData = TRUE)
}
