% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment Test Results}
\usage{
plotEnrichment(
  object,
  modes = NULL,
  onlyCase = TRUE,
  onlyPOS = FALSE,
  splitby = c("none", "verdict", "condType"),
  returnData = FALSE
)
}
\arguments{
\item{object}{The tibble object obtained from running \code{featureEnrich}.}

\item{modes}{Which modes to include in plot? If empty, all modes included.}

\item{onlyCase}{If TRUE, only "case" predicted samples included. Default: TRUE.}

\item{onlyPOS}{If TRUE, only "POS" labeled samples included. Default: FALSE.}

\item{splitby}{Metadata by which to split plots. Can be "none", "verdict", or "condType".}

\item{returnData}{If TRUE, plot data is returned instead of plotting. Default: FALSE}
}
\value{
A named list of ggplot objects.
}
\description{
Plot Enrichment Test Results
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Run RLSeq
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")
rlr <- RLSeq(rlr)

# Plot enrichment
plotEnrichment(rlr)

# split by verdict
plotEnrichment(rlr, onlyCase=FALSE, splitby="verdict")

# split by "condType" and only return plotting data
plotEnrichment(rlr, onlyCase=FALSE, splitby="condType", returnData=TRUE)

# Only include DRIP-family
plotEnrichment(rlr, modes=c("DRIP", "DRIPc", "qDRIP", "sDRIP", "ssDRIP"),
               onlyCase=FALSE, splitby="verdict")

}
}
