% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFingerprint}
\alias{plotFingerprint}
\title{Plot noise analysis results as a fingerprint plot}
\usage{
plotFingerprint(object)
}
\arguments{
\item{object}{An RLRanges object with \link{noiseAnalyze} already run.}
}
\value{
A ggplot object. See also \link[ggplot2:ggplot]{ggplot2::ggplot}.
\enumerate{
\item \code{noiseComparisonPlot}
}
\itemize{
\item A plot showing the noise analysis
results from the user-supplied sample compared to similar samples from
RLBase.
}
}
\description{
Plots the results of the noise analysis in \link{noiseAnalyze}. Creates a
fingerprint plot like those developed by
\href{https://pubmed.ncbi.nlm.nih.gov/22499706/}{Diaz et al, 2012} and those
provided by
\href{https://deeptools.readthedocs.io/en/develop/content/tools/plotFingerprint.html}{deepTools}.
}
\details{
The term "Fingerprint plot" comes from deepTools.
}
\examples{

# Example RLRanges dataset with analyzeRLFS() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Plot RLFS res
plotFingerprint(rlr)

}
