% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLFSRes}
\alias{plotRLFSRes}
\title{Plot RLFS analysis results}
\usage{
plotRLFSRes(object, plotName = NULL, fft = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object with \link{analyzeRLFS} already run.
Alternatively, can be the RLFS results object from an RLRanges (from
\code{rlresult(object, "rlfsRes")}).}

\item{plotName}{A Sample name used for plotting.
If blank, the RLRanges \code{sampleName} metadata entry is used (see \link{RLRanges}).}

\item{fft}{If TRUE, the Fourier transform of the Z-score is plotted instead.
Default: FALSE.}

\item{...}{Additional parameters passed to \link[ggplot2:ggplot]{ggplot2::ggplot}.}
}
\value{
A ggplot object. See also \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Plots the results of the R-loop-forming sequences (RLFS) analysis. The plot is a
metaplot of the Z score distribution around RLFS with the p value from
permutation testing annotated. See also \link{analyzeRLFS}.
}
\examples{

# Example RLRanges dataset with analyzeRLFS() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Plot RLFS res
plotRLFSRes(rlr)

# Plot the Fourier transform instead
plotRLFSRes(rlr, fft = TRUE)
}
