% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLRegionOverlap}
\alias{plotRLRegionOverlap}
\title{Plot RL-Region overlap with RLRanges}
\usage{
plotRLRegionOverlap(object, returnData = FALSE)
}
\arguments{
\item{object}{An RLRanges object with \code{rlRegionTest()} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting. Default: FALSE}
}
\value{
A venn diagram ggplot object.
}
\description{
Plot RL-Region overlap with RLRanges
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Run RLSeq
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")
rlr <- RLSeq(rlr)

# Plot RL-Region overlap
plotRLRegionOverlap(rlr)

# Return data only
plotRLRegionOverlap(rlr, returnData=TRUE)

}
}
