% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTxFeatureOverlap}
\alias{plotTxFeatureOverlap}
\title{Plot Transcript Feature Overlap}
\usage{
plotTxFeatureOverlap(object, mode = "auto", returnData = FALSE)
}
\arguments{
\item{object}{An RLRanges object with \link{txFeatureOverlap} already run.}

\item{mode}{A \code{character} containing the R-loop data mode to compare
against. See details for more information.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object or a \code{tbl} if \code{returnData} is \code{TRUE}.
}
\description{
Plots the results of \link{txFeatureOverlap} alongside the average from
public R-loop datasets. This allows comparison of user-supplied
samples with data that is expected to be simialr.
}
\details{
\subsection{Mode}{

The \code{mode} parameter specifies the R-loop modality to compare the
user-supplied sample against in the plot. The default, \code{"auto"}
specifies that the \code{mode} from the supplied \code{RLRanges} object will
be used. Only one mode can be specified. For a list of applicable modes,
see \code{auxdata$available_modes}.
}

\subsection{Plot}{

The plot is a stacked bar chart showing the proportion of peaks overlapping
transcript features for the supplied \code{RLRanges} object. Additionally, the
average of the \link{txFeatureOverlap} analysis for all samples within the
specified modes are also shown as a background comparison.

This style of analysis enables a user to see the transcript features
overlapping their peaks and compare those results to the average within
relevant public datasets.
}
}
\examples{

# Example dataset with rlRegionTest() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Plot RL-Region overlap
plotTxFeatureOverlap(rlr)

# Return data only
plotRLRegionOverlap(rlr, returnData = TRUE)
}
