% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepFeatures}
\alias{prepFeatures}
\title{A Caret model used for preprocessing datasets in predictCondition().}
\format{
An object of class \code{preProcess} of length 21.
}
\usage{
prepFeatures
}
\description{
A Caret model used for preprocessing datasets in predictCondition().
}
\keyword{datasets}
