% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomWindows}
\alias{randomWindows}
\title{Random Genomic Windows}
\format{
An object of class \code{list} of length 3.
}
\usage{
randomWindows
}
\description{
A collection of random genomic windows for use with \link{noiseAnalyze}. See the
\code{data-raw/genome_masks.R} script for the processing steps.
}
\details{
\subsection{Structure}{

\code{randomWindows} is a named list of \code{tbl} objects containing ~1000 random
genomic regions. The names are the genomes to which the regions correspond.

Columns (mirrors
\href{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}{bed3 format}).
\enumerate{
\item \code{chrom} - the chromosome in UCSC format
\item \code{start} - the starting position
\item \code{end} - the end position
}
}
}
\examples{
randomWindows

}
\keyword{datasets}
