% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate an RLSeq Report}
\usage{
report(object, reportPath = "rlreport.html", quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object.}

\item{reportPath}{A path indicating the report output HTML file. Default: "rlreport.html"}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{rmarkdown::render()}}
}
\value{
TRUE
}
\description{
Generate an RLSeq Report
}
\examples{

rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")
rlr <- RLRanges(pks, coverage=cvg,  genome="hg38", mode="DRIP")
rlr <- RLSeq(rlr)

report(rlr)

}
