% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rlbaseNoiseAnalyze}
\alias{rlbaseNoiseAnalyze}
\title{RLBase Sample Noise Analysis Results}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 704 rows and 2 columns.
}
\usage{
rlbaseNoiseAnalyze
}
\description{
Average signal from noise analysis of RLBase samples. See \link{noiseAnalyze} for
more detail regarding how signal was initially calculated.
}
\details{
\subsection{Structure}{

\code{rlbaseNoiseAnalyze} is a \code{tbl} with the following columns:
\itemize{
\item \code{rlsample}
\itemize{
\item The RLBase sample identifier for the sample.
\item Matches the \code{rlsample} column in \link[RLHub:rlbase_samples]{RLHub::rlbase_samples}
}
\item \code{value}
\itemize{
\item The mean signal from noise analysis. See \link{noiseAnalyze}.
}
}
}
}
\examples{
rlbaseNoiseAnalyze

}
\keyword{datasets}
