% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RLRanges-class}
\alias{RLRanges-class}
\alias{RLRanges}
\title{Construct RLRanges Dataset}
\usage{
RLRanges(
  peaks = GenomicRanges::GRanges(),
  coverage = character(1),
  genome = character(1),
  mode = character(1),
  condType = character(1),
  sampleName = "User-selected sample",
  quiet = FALSE
)
}
\arguments{
\item{peaks}{Path/URL to peak file or a GRanges object.}

\item{coverage}{Path/URL to bigWig file. If not supplied, correlation tests will be skipped.}

\item{genome}{UCSC genome ID. Acceptable types are listed in RLSeq::aux$available_genomes.}

\item{mode}{Type of R-loop mapping from which peaks and coverage were
derived. Acceptable types are listed in RLSeq::aux$available_modes$mode. Can
be unspecified.}

\item{condType}{One of "POS" (e.g., S9.6 -RNH1), "NEG" (e.g., S9.6 +RNH1),
or "NULL" (e.g., Input control.). Can be unspecified.}

\item{sampleName}{A unique name for identifying this sample. Can be unspecified.}

\item{quiet}{If TRUE, messages and warnings are suppressed. Default: FALSE.}
}
\description{
\code{RLRanges} is a subclass of \code{GRanges}, which stores R-loop peaks
and metadata about the R-loop-mapping experiment.
}
\examples{

pks <- "https://rlbase-data.s3.amazonaws.com/peaks/SRX1025890_hg38.broadPeak"
cvg <- "https://rlbase-data.s3.amazonaws.com/coverage/SRX1025890_hg38.bw"

rlrs <- RLRanges(
  peaks = pks,
  coverage = cvg,
  mode = "DRIP",
  genome = "hg38",
  condType = "POS",
  sampleName = "TC32 DRIP-Seq"
)
}
