% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRangesFromRLBase.R
\name{RLRangesFromRLBase}
\alias{RLRangesFromRLBase}
\title{Access RLBase samples as RLRanges}
\usage{
RLRangesFromRLBase(acc, quiet = FALSE)
}
\arguments{
\item{acc}{The sample ID of the RLBase object. \code{rlsample} column in \code{RLHub::rlbase_samples()}.}

\item{quiet}{If TRUE, messages will be suppressed. Default: FALSE}
}
\value{
An RLRanges object.
}
\description{
Accessor function which returns any sample in RLBase as an RLRanges object for
use with RLSeq.
}
\examples{

rlr <- RLRangesFromRLBase("SRX1070676")

}
