% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rlRegions}
\alias{rlRegions}
\title{A Tibble containing the R-loop regions identified through metanalysis of publicly-available
R-loop mapping samples (as described in Miller et al. 2021). The tibble includes additional
information, such as the RLoop Region ID, the number of studies, samples, and modes
that RL Region is found in, the average signal and qVal of peaks overlapping with it,
any genes which overlap with it, the correlation with expression corrR and corrPval,
and whether the RLoop exists in Repeat regions or RLFS regions.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61726 rows and 11 columns.
}
\usage{
rlRegions
}
\description{
A Tibble containing the R-loop regions identified through metanalysis of publicly-available
R-loop mapping samples (as described in Miller et al. 2021). The tibble includes additional
information, such as the RLoop Region ID, the number of studies, samples, and modes
that RL Region is found in, the average signal and qVal of peaks overlapping with it,
any genes which overlap with it, the correlation with expression corrR and corrPval,
and whether the RLoop exists in Repeat regions or RLFS regions.
}
\keyword{datasets}
