% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlRegionTest.R
\name{rlRegionTest}
\alias{rlRegionTest}
\title{R-Loop region test}
\usage{
rlRegionTest(object)
}
\arguments{
\item{object}{An RLRanges object with genome "hg38".}
}
\value{
An RLRanges object with test results included.
}
\description{
Tests the overlap of user-supplied ranges with R-loop regions.
}
\examples{
\dontrun{

# Get example data
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")

# RL Region Test
rlRegionTest(rlr)

}
}
