% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{rlresult}
\alias{rlresult}
\title{Result accessor function}
\usage{
rlresult(object, resultName)
}
\arguments{
\item{object}{RLRanges object.}

\item{resultName}{Name of the result slot to access.}
}
\description{
Result accessor function
}
