% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{rlresult}
\alias{rlresult}
\alias{RLResults}
\alias{RLResults-class}
\title{RLSeq Results}
\usage{
rlresult(object, resultName)
}
\arguments{
\item{object}{\link{RLRanges} object.}

\item{resultName}{Name of the result slot to access. See \emph{details}.}
}
\value{
The contents of the requested slot.
}
\description{
Functions for creating and accessing the R-loop results (RL Results). These
are a type of object used for holding the results of the tests implemented
in RLSeq. They can be accessed using the \code{rlresult} function.
}
\details{
\subsection{Slot descriptions}{
\itemize{
\item \code{featureEnrichment}
\itemize{
\item The \code{tbl} generated from running \link{featureEnrich}.
\item The structure and column descriptions are provided in detail within
\link[RLHub:featureEnrichment]{RLHub::feat_enrich_samples}.
}
\item \code{correlationMat}
\itemize{
\item The \code{matrix} generated from running \link{corrAnalyze}.
\item Contains pairwise pearson correlations between all samples in
\href{https://gccri.bishop-lab.uthscsa.edu/rlbase/}{RLBase}
and the supplied RLRanges object.
}
\item \code{rlfsRes}
\itemize{
\item The \code{list} generated from running \link{analyzeRLFS}.
\item See \link{analyzeRLFS} for description of structure.
}
\item \code{noiseAnalysis}
\itemize{
\item The \code{tbl} generated from running \link{noiseAnalyze}.
}
\item \code{txFeatureOverlap}
\itemize{
\item The \code{tbl} generated from running \link{txFeatureOverlap}.
}
\item \code{geneAnnoRes}
\itemize{
\item The \code{tbl} generated from running \link{geneAnnotation}.
}
\item \code{predictRes}
\itemize{
\item The \code{list} generated from running \link{predictCondition}.
}
\item \code{rlRegionRes}
\itemize{
\item The \code{list} generated from running \link{rlRegionTest}.
}
}
}
}
\examples{

rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

rlresult(rlr, "predictRes")
}
