% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rlsampleTxOl}
\alias{rlsampleTxOl}
\title{RLBase Sample Transcript Feature Overlaps}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4754 rows and 4 columns.
}
\usage{
rlsampleTxOl
}
\description{
Summary statistics from transcript feature overlap analysis of peaks
from all RLBase samples.
}
\details{
\subsection{Structure}{

\code{rlsampleTxOl} is a \code{tbl} with the following columns:
\itemize{
\item \code{rlsample}
\itemize{
\item The RLBase sample identifier for the sample.
\item Matches the \code{rlsample} column in \link[RLHub:rlbase_samples]{RLHub::rlbase_samples}
}
\item \code{feature}
\itemize{
\item The transcript feature for which overlap analysis was performed.
\item These features were derived from the Transcript Features collection
described in \link[RLHub:annotations]{RLHub::annotations}
}
\item \code{n}
\itemize{
\item The raw number of peaks from the sample overlapping a feature.
}
\item \code{pct}
\itemize{
\item The proportion of peaks fro the sample overlapping a feature.
}
}
}
}
\examples{
rlsampleTxOl

}
\keyword{datasets}
