% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLSeq.R
\name{RLSeq}
\alias{RLSeq}
\title{Run RLSeq}
\usage{
RLSeq(object, quiet = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}
}
\value{
An RLRanges object with all results available.
}
\description{
Runs the full RLSeq pipeline.
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Build RLRanges object
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")

# Run RLSeq
rlr <- RLSeq(rlr)

}
}
