% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmapSamps}
\alias{rmapSamps}
\title{A tibble containing information about the samples in RMapDB.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 487 rows and 13 columns.
}
\usage{
rmapSamps
}
\description{
A tibble containing information about the samples in RMapDB.
}
\keyword{datasets}
