% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRX1025890}
\alias{SRX1025890}
\title{Results of running RSeq on SRX1025890}
\format{
An object of class \code{list} of length 9.
}
\usage{
SRX1025890
}
\description{
Results of running RSeq on SRX1025890
}
\keyword{datasets}
