% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRX1025890_peaks}
\alias{SRX1025890_peaks}
\title{GRanges. Peaks from running RSeq-CLI on SRX1025890 - in "hg38" format.}
\format{
An object of class \code{GRanges} of length 2902.
}
\usage{
SRX1025890_peaks
}
\description{
GRanges. Peaks from running RSeq-CLI on SRX1025890 - in "hg38" format.
}
\keyword{datasets}
