% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRX1025890_peaks_hg19}
\alias{SRX1025890_peaks_hg19}
\title{GRanges. Peaks from running liftUtil(SRX1035890_peaks, "hg38", "hg19") in "hg19" format.}
\format{
An object of class \code{GRanges} of length 2895.
}
\usage{
SRX1025890_peaks_hg19
}
\description{
GRanges. Peaks from running liftUtil(SRX1035890_peaks, "hg38", "hg19") in "hg19" format.
}
\keyword{datasets}
