% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tableToRegions}
\alias{tableToRegions}
\title{Table to Regions}
\usage{
tableToRegions(table)
}
\arguments{
\item{table}{A tibble in "Table" format from RLHub.}
}
\value{
A tibble in "regions" format.
}
\description{
Helper function to Convert "table" format to "regions" format.
}
