% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBed.R
\name{toBed}
\alias{toBed}
\title{To Bed}
\usage{
toBed(granges, write = FALSE, filename = NULL)
}
\arguments{
\item{granges}{A GRanges object containing DRIP-Seq peaks}

\item{write}{A boolean determining if the converted DataFrame will be written to a .bed file}

\item{filename}{A string containing the desired file name if writing to file}
}
\value{
A DataFrame object containing the GRanges content formatted according to .bed standards
}
\description{
Converts a GRanges object to a .bed formatted DataFrame and writes to file if requested
}
