% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txFeatureOverlap.R
\name{txFeatureOverlap}
\alias{txFeatureOverlap}
\title{Calculate overlap with transcript features}
\usage{
txFeatureOverlap(object, quiet = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{quiet}{If TRUE, messages will be suppressed. Default: FALSE}
}
\value{
An RLRanges object containing the results of the enrichment test
accessed via \code{rlresult(object, "txFeatureOverlap")}. The results
are in \code{tbl} format.
}
\description{
Tests the overlap of transcript features with supplied peaks. See \emph{details}.
}
\details{
\subsection{Method}{

Transcript annotations were curated as part of the
\href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} workflow
and are provided via \link[RLHub:annotations]{RLHub::annotations}.

In \code{txFeatureOverlap}, each annotation "type" (e.g., "Exons", "Introns", etc)
is compared to the supplied RLRanges, yielding overlap statistics with
the following procedure:
\enumerate{
\item For each annotation type, the peaks are overlapped with the annotations.
\item Then the number of overlapping peaks is counted and summarised using a
priority order. This order determines which feature is assigned to a peak
when that peak overlaps multiple features. The order is "TSS", "TTS", "5'UTR",
"3'UTR", "Exon", "Intron", "Intergenic".
}
}
}
\examples{

# Example RLRanges dataset
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# RL Region Test
txFeatureOverlap(rlr)

}
