% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRLFS.R
\name{analyzeRLFS}
\alias{analyzeRLFS}
\title{Analyze RLFS}
\usage{
analyzeRLFS(object, mask = NULL, quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object.}

\item{mask}{GRanges object containing masked genomic ranges.
Not needed unless masked genome unavailable (see genomeMasks).
Custom masks can be generated using regioneR::getMask().}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{regioneR::permTest()}}
}
\value{
An RLRanges object with RLFS analysis results included.
}
\description{
Analyzes the enrichment of ranges within R-loop forming sequences (RLFS).
}
\examples{

# Example dataset
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Perform RLFS analysis
rlr <- analyzeRLFS(rlr)
}
