% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRLFS.R
\name{analyzeRLFS}
\alias{analyzeRLFS}
\title{Analyze RLFS}
\usage{
analyzeRLFS(
  object,
  mask = NULL,
  quiet = FALSE,
  useMask = TRUE,
  noZ = FALSE,
  ntimes = 100,
  stepsize = 50,
  ...
)
}
\arguments{
\item{object}{An RLRanges object.}

\item{mask}{GRanges object containing masked genomic ranges.
Not needed unless masked genome unavailable (see \code{RLSeq:::genomeMasks}).
Custom masks can be generated using regioneR::getMask().}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{useMask}{If FALSE, masked genome is not used. This is
not recommended unless a mask is unavailable as it can lead to spurious
results. Default: TRUE.}

\item{noZ}{If TRUE, Z-score distribution is not calculated. Default: FALSE.}

\item{ntimes}{Number of permutations to perform (default: 100).}

\item{stepsize}{The step size for calculating the Z score distribution.
Default: 50. See also \code{\link[regioneR:localZScore]{regioneR::localZScore()}}.}

\item{...}{Arguments passed to \code{\link[regioneR:permTest]{regioneR::permTest()}}.}
}
\value{
An RLRanges object with RLFS analysis results included.
}
\description{
Analyzes the enrichment of ranges within R-loop forming sequences (RLFS).
}
\examples{

# Example dataset
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Perform RLFS analysis (remove ntimes=2 and noZ=TRUE for a typical analysis)
rlr <- analyzeRLFS(rlr, ntimes = 2, noZ = TRUE)
}
