% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRLFS.R
\name{analyzeRLFS}
\alias{analyzeRLFS}
\title{Analyze RLFS}
\usage{
analyzeRLFS(object, mask = NULL, quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object.}

\item{mask}{GRanges object containing masked genomic ranges.
Not needed unless masked genome unavailable (see RLSeq::genomeMasks).}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{regioneR::permTest()}}
}
\value{
An RLRanges object with RLFS analysis results included.
}
\description{
Analyzes the enrichment of peaks within R-loop forming sequences.
}
\examples{

rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
rlr <- RLRanges(pks, genome = "hg38", mode = "DRIP")
rlr <- analyzeRLFS(rlr)
}
