% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRLFS.R
\name{analyzeRLFS}
\alias{analyzeRLFS}
\title{Analyze RLFS}
\usage{
analyzeRLFS(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object generated with \code{RLRanges()}. Required.}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{regioneR::permTest()}}

\item{mask}{GRanges object containing masked genomic ranges.
Not needed unless masked genome unavailable (see RLSeq::genomeMasks).}
}
\value{
A named list containing the results of RegioneR,
peaks annotated with the RLFS they overlap with,
and the Z score results within 3000 BP of an RLFS.
}
\description{
Analyzes the enrichment of peaks within R-loop forming sequences.
}
\examples{

pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
rlr <- RLRanges(pks, genome="hg38", mode="DRIP")
rlr <- analyzeRLFS(rlr)

}
