% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRLFS.R
\name{analyzeRLFS}
\alias{analyzeRLFS}
\title{Analyze RLFS}
\usage{
analyzeRLFS(peaks, genome = "hg38", chrom_sizes = NULL, RLFS = NULL, ...)
}
\arguments{
\item{peaks}{A GRanges object containing the R-loop ranges to check.}

\item{genome}{UCSC genome which peaks were generated from. Ignored if "chrom_sizes" or "RLFS" is specified.}

\item{chrom_sizes}{A data.frame containing two columns with chromosome names
and chromosome sizes.}

\item{RLFS}{A GRanges object containing the R-loop forming sequences to
compare against.}

\item{...}{Arguments passed to \code{RegioneR::permTest()}}
}
\value{
A named list containing the results of RegioneR,
peaks annotated with the RLFS they overlap with,
and the Z score results within 3000 BP of an RLFS.
}
\description{
Analyzes the enrichment of peaks within R-loop forming sequences.
}
\examples{

result <- RSeqR::analyzeRLFS(RSeqR::SRX1025890_peaks, genome="hg38")

}
