% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotate Peaks}
\usage{
annotatePeaks(bed)
}
\arguments{
\item{peaks}{A .bed file containing DRIP-Seq peaks}
}
\value{
A .bed file containing annotated DRIP-Seq peaks
}
\description{
Annotates DRIP-Seq peaks with gene-level information
}
