% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_genomes}
\alias{available_genomes}
\title{data.frame. Contains information about the genes available in UCSC database.
Result of running RLSeq:::buildAvailableGenomes()}
\format{
An object of class \code{data.frame} with 199 rows and 27 columns.
}
\usage{
available_genomes
}
\description{
data.frame. Contains information about the genes available in UCSC database.
Result of running RLSeq:::buildAvailableGenomes()
}
\keyword{datasets}
