% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDatasets.R, R/utils.R
\name{buildAvailableGenomes}
\alias{buildAvailableGenomes}
\title{Build Available Genomes
Helper Function which builds the UCSC available genomes dataset}
\usage{
buildAvailableGenomes(test = FALSE, ...)

buildAvailableGenomes(test = FALSE, ...)
}
\arguments{
\item{test}{if TRUE, will only build info set for first 2 genomes in UCSC.}

\item{...}{arguments passed to \code{getEffectiveGenomeSizes()}}
}
\value{
A data frame of avaialable genomes and corresponding info.

A data frame of avaialable genomes and corresponding info.
}
\description{
Build Available Genomes
Helper Function which builds the UCSC available genomes dataset

Build Available Genomes
Helper Function which builds the UCSC available genomes dataset
}
