% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAnalyze.R
\name{corrAnalyze}
\alias{corrAnalyze}
\title{Analyze Correlations}
\usage{
corrAnalyze(coverage, genome = "hg38")
}
\arguments{
\item{coverage}{The results list from running analyzeRLFS().}

\item{genome}{The UCSC genome ID to use. (Currently only "hg38" is supported)}
}
\value{
A named list containing the results of correlation analysis.
}
\description{
Finds the pairwise correlation in signal around gold-standard R-Loop sites
between the query sample and the RMapDB database.
}
\examples{

URL <- "https://rmapdb-data.s3.us-east-2.amazonaws.com/bigwigs/rseq-coverage-unstranded/SRX1025890_TC32_NT_DRIP.hg38.bw"
BW_FILE <- "SRX1025890.bw"
download.file(URL, destfile=BW_FILE)
result <- RSeqR::corrAnalyze(BW_FILE, genome="hg38")

}
