% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAnalyze.R
\name{corrAnalyze}
\alias{corrAnalyze}
\title{Analyze Correlations}
\usage{
corrAnalyze(object, force = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{force}{Force corrAnalyze() to run, even if on Windows. Default: FALSE.}
}
\value{
An RLRanges object with correlation results included.
}
\description{
Finds the pairwise correlation in signal around gold-standard R-Loop sites
between the query sample and the RLBase database.
}
\details{
Currently, this does not work on windows.
}
\examples{

# Example RLRanges object
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# corrAnalyze does not work on Windows OS
if (.Platform$OS.type != "windows") {

    # run corrAnalyze
    rlr <- corrAnalyze(rlr)
    
}

}
