% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAnalyze.R
\name{corrAnalyze}
\alias{corrAnalyze}
\title{Analyze Correlations}
\usage{
corrAnalyze(object)
}
\arguments{
\item{object}{An RLRanges object.}
}
\value{
An RLRanges object with correlation results included.
}
\description{
Finds the pairwise correlation in signal around gold-standard R-Loop sites
between the query sample and the RMapDB database.
}
\examples{

pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")

rlr <- corrAnalyze(rlr)
}
