% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAnalyze.R
\name{corrAnalyze}
\alias{corrAnalyze}
\title{Analyze Correlations}
\usage{
corrAnalyze(object)
}
\arguments{
\item{object}{An RLRanges object.}
}
\value{
An RLRanges object with correlation results included.
}
\description{
Finds the pairwise correlation in signal around gold-standard R-Loop sites
between the query sample and the RLBase database.
}
\examples{

# Example RLRanges object
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# run corrAnalyze
rlr <- corrAnalyze(rlr)
}
