% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Plot Correlation Results}
\usage{
corrHeatmap(object, returnData = FALSE, complex = TRUE, ...)
}
\arguments{
\item{object}{An RLRanges with \code{corrAnalyze()} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}

\item{complex}{If TRUE, ComplexHeatmap will be used for plotting.
Otherwise, pheatmap. Default: TRUE}

\item{...}{For internal use.}
}
\value{
A Heatmap object.
}
\description{
Plot Correlation Results
}
\examples{

# Example RLRanges data with corrAnalyze() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Corr heatmap
corrHeatmap(rlr)
}
