% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Plot Correlation Results}
\usage{
corrHeatmap(object, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run.}

\item{...}{For internal use.}
}
\value{
A Heatmap object.
}
\description{
Plot Correlation Results
}
