% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Plot Correlation Results}
\usage{
corrHeatmap(object, returnData = FALSE, complex = TRUE, ...)
}
\arguments{
\item{object}{An RLRanges with \link{corrAnalyze} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}

\item{complex}{If TRUE, \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap} will be used for plotting.
Otherwise, \link[pheatmap:pheatmap]{pheatmap::pheatmap} is used. Default: TRUE}

\item{...}{For internal use.}
}
\value{
A plot object or plotting data (if \code{returnData} is \code{TRUE}).
}
\description{
Plots a heatmap to visualize the pairwise Pearson correlation matrix
generated via \link{corrAnalyze}.
}
\examples{

# Example RLRanges data with corrAnalyze() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Corr heatmap
corrHeatmap(rlr)
}
