% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Plot Correlation Results}
\usage{
corrHeatmap(object, returnData = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting. Default: FALSE}

\item{...}{For internal use.}
}
\value{
A Heatmap object.
}
\description{
Plot Correlation Results
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Run RLSeq
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")
rlr <- RLSeq(rlr)

# Corr heatmap
corrHeatmap(rlr)
corrHeatmap(rlr, returnData=TRUE)
}
}
