% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrich.R
\name{featureEnrich}
\alias{featureEnrich}
\title{Tests Genomic Feature Enrichment}
\usage{
featureEnrich(peaks, genome = c("hg38", "mm10"))
}
\arguments{
\item{peaks}{A GRanges object containing R-loop peaks}

\item{genome}{UCSC genome identifier to use. Can be "hg38" or "mm10" currently.}
}
\value{
A tibble containing each annotation, the Log2 fold change of observed
overlap vs expected, and the p.adjusted value.
}
\description{
Tests the enrichment of genomic features in supplied R-loop peaks.
}
\examples{

RSeqR::featureEnrich(RSeqR::SRX1025890_peaks, genome="hg38")

}
