% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrich.R
\name{featureEnrich}
\alias{featureEnrich}
\title{Test Genomic Feature Enrichment}
\usage{
featureEnrich(
  object,
  annotype = c("primary", "full"),
  annotations = NULL,
  downsample = 10000,
  quiet = FALSE
)
}
\arguments{
\item{object}{An RLRanges object.}

\item{annotype}{The type of annotations to use. Can be one of "primary"
or "full". Default: "primary". See
\link[RLHub:annotations]{RLHub::annotations} for greater detail.}

\item{annotations}{A custom annotation list of the same structure
described in \link[RLHub:annotations]{RLHub::annotations}.}

\item{downsample}{If a numeric, data will be down sampled to the requested
number of peaks. This improves the speed of genomic shuffling and
helps prevent p-value inflation.
If FALSE, then downsampling will not be performed. Default: 10000.}

\item{quiet}{If TRUE, messages will be suppressed. Default: FALSE}
}
\value{
An RLRanges object containing the results of the enrichment test
accessed via \code{rlresult(object, "featureEnrichment")}. The results
are in \code{tbl} format. For a full description of all columns in the output
table see \link[RLHub:featureEnrichment]{RLHub::feat_enrich_samples}.
}
\description{
Tests the enrichment of genomic features in supplied peaks. See \emph{details}.
}
\details{
\subsection{Method}{

Annotations relevant to R-loops were curated as part of the
\href{https://github.com/Bishop-Laboratory/RLBase-data}{RLBase-data} workflow
and are provided via \link[RLHub:annotations]{RLHub::annotations}.

In \code{featureEnrich}, each annotation "type" (e.g., "Exons", "Introns", etc)
is compared to the supplied RLRanges, yielding enrichment statistics with
the following procedure:
\enumerate{
\item For each annotation type, the peaks are overlapped with the annotations.
\item Then, \link[valr:bed_reldist]{valr::bed_reldist} is used to find the relative distance
distribution between the peaks and the annotations for both the supplied
RLRanges and shuffled RLRanges (via \link[valr:bed_shuffle]{valr::bed_shuffle}).
Significance of the relative distance is calculated via \link[stats:ks.test]{stats::ks.test}.
\item Then, Fisher’s exact test is implemented via \link[valr:bed_fisher]{valr::bed_fisher}
to obtain the significance of the overlap and the odds ratio.
}
}
}
\examples{

# Example RLRanges dataset
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# RL Region Test
featureEnrich(rlr)

# With custom annotations
small_anno <- list(
    "Centromeres" = readr::read_csv(
        system.file("extdata", "Centromeres.csv.gz", package = "RLSeq"),
        show_col_types = FALSE
    )
)
featureEnrich(rlr, annotations = small_anno)
}
