% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Annotate R-Loops with Genes}
\usage{
geneAnnotation(object, txdb = NULL, quiet = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{txdb}{The TxDb or EnsDb object containing gene annotations. If not
supplied, annotations will be automatically downloaded from AnnotationHub.}

\item{quiet}{If TRUE, messages and warnings are suppressed.}
}
\value{
An RLRanges object with annotations added.
}
\description{
Annotates R-loop peaks as a GRanges object with gene-level information
}
\examples{

# Example RLRanges data
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Perform gene annotation
rlr <- geneAnnotation(rlr)

# Supply custom TxDb if needed
if (GenomeInfoDb::genome(rlr)[1] == "hg19") {
    library(TxDb.Hsapiens.UCSC.hg19.knownGene)
    rlr <- geneAnnotation(rlr, txdb = TxDb.Hsapiens.UCSC.hg19.knownGene)
}
}
