% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Annotate R-Loops with Genes}
\usage{
geneAnnotation(object, txdb = NULL)
}
\arguments{
\item{object}{An RLRanges object.}

\item{txdb}{The TxDb or EnsDb object containing gene annotations. If not
supplied, annotations will be automatically downloaded from AnnotationHub.}
}
\value{
An RLRanges object with genes overlaps included.
}
\description{
Annotates RLRanges with entrez ids for overlapping genes.
}
\examples{

# Example RLRanges data
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Perform gene annotation
rlr <- geneAnnotation(rlr)

# Supply custom TxDb if needed
if (GenomeInfoDb::genome(rlr)[1] == "hg19") {
    library(TxDb.Hsapiens.UCSC.hg19.knownGene)
    rlr <- geneAnnotation(rlr, txdb = TxDb.Hsapiens.UCSC.hg19.knownGene)
}
}
