% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Annotate R-Loops with Genes}
\usage{
geneAnnotation(object, txdb = NULL, quiet = FALSE)
}
\arguments{
\item{object}{An RLRanges object.}

\item{txdb}{The TxDb or EnsDb object containing gene annotations. If not
supplied, annotations will be automatically downloaded from AnnotationHub.}

\item{quiet}{If TRUE, messages and warnings are suppressed.}
}
\value{
An RLRanges object with annotations added.
}
\description{
Annotates R-loop peaks as a GRanges object with gene-level information
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Get RLRanges object
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")

# Perform gene annotation
rlr <- geneAnnotation(rlr)

# Supply custom TxDb if needed
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
rlr <- geneAnnotation(rlr, txdb=TxDb.Hsapiens.UCSC.hg19.knownGene)

}
}
