% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Annotate R-Loops with Genes}
\usage{
geneAnnotation(peaks, genome)
}
\arguments{
\item{peaks}{A GRanges object containing R-loop peaks}

\item{genome}{UCSC genome which peaks were generated from.
Only "hg38" and "mm10" currently available.
Use RSeqR::liftUtil() to convert to the correct format, if needed. Additionally,
this can be an EnsDb or TxDb object.}
}
\value{
A tibble object containing annotated R-loop peaks
}
\description{
Annotates R-loop peaks as a GRanges object with gene-level information
}
\examples{

geneAnnotation(RSeqR::SRX1025890_peaks)

}
