% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genomeMasks}
\alias{genomeMasks}
\title{Genome Masks}
\format{
An object of class \code{list} of length 8.
}
\usage{
genomeMasks
}
\description{
A collection of genome masks for use with \link{analyzeRLFS}. See the
\code{data-raw/genome_masks.R} script for the processing steps.
}
\details{
\subsection{Structure}{

\code{genomeMasks} is a named list of \code{GRanges} objects. Each entry in the
list follows the naming convention: \verb{<genome>.masked}, where \verb{<genome>}
is an official UCSC genome ID. Each entry contains
a \code{GRanges} object with the masked ranges from \verb{<genome>}. The genomes
provided correspond to the masked genomes available in
\link[BSgenome:available.genomes]{BSgenome::available.genomes}.
}
}
\examples{
genomeMasks

}
\keyword{datasets}
