% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSizes}
\alias{getChromSizes}
\title{Get Chrom Sizes
Helper function which extracts chrom sizes from an RLRanges object.}
\usage{
getChromSizes(object)
}
\arguments{
\item{object}{An RLRanges object.}
}
\value{
A tibble containing chrom sizes
}
\description{
Get Chrom Sizes
Helper function which extracts chrom sizes from an RLRanges object.
}
