% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGSSignal}
\alias{getGSSignal}
\title{Get GS Signal}
\usage{
getGSSignal(coverage, gssignal)
}
\arguments{
\item{coverage}{The path to a .bigWig file (can be a URL)}

\item{gssignal}{The GS signal obtained from RLHub.}
}
\value{
A named list containing the results of correlation analysis.
}
\description{
Extract signal around GS R-loop sites
}
