% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRLFSAnno}
\alias{getRLFSAnno}
\title{Get RLFS Anno
Helper function that retrieves RLFS}
\usage{
getRLFSAnno(object)
}
\arguments{
\item{object}{An RLRanges object.}
}
\value{
A GRanges object with RLFS for that species.
}
\description{
Get RLFS Anno
Helper function that retrieves RLFS
}
