% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsSignalRMapDB}
\alias{gsSignalRMapDB}
\title{The matrix generated by summarizing all hg38 RMapDB bigWig files around the
"Gold-Standard" R-loop sites defined by Chedin et al.
https://doi.org/10.15252/embj.2020106394.
Columns are sample names and rows are hg38 genomic coordinates.
See data-raw/gsSignalRMapDB.R for processing steps.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3452 rows and 301 columns.
}
\usage{
gsSignalRMapDB
}
\description{
The matrix generated by summarizing all hg38 RMapDB bigWig files around the
"Gold-Standard" R-loop sites defined by Chedin et al.
https://doi.org/10.15252/embj.2020106394.
Columns are sample names and rows are hg38 genomic coordinates.
See data-raw/gsSignalRMapDB.R for processing steps.
}
\keyword{datasets}
