% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{makeReport}
\alias{makeReport}
\title{Generate report}
\usage{
makeReport(data, output_file = "RSeqReport.html", ...)
}
\arguments{
\item{data}{A named list containing \code{annotated_peaks}, \code{feature_overlaps},
and \code{rlfs_results}}

\item{...}{Arguments passed to \code{rmarkdown::render()}}

\item{outputFile}{A path indicating the report output HTML file.}
}
\description{
Generates RSeqR Quality Report
}
\examples{

RSeqR::makeReport(RSeqR::SRX1025890, output_file = "report.html")

}
