% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment Test Results}
\usage{
plotEnrichment(sampleRes, rlbaseRes = NULL, yval = "stat_fisher_rl")
}
\arguments{
\item{sampleRes}{The tibble obejct obtained from running \code{featureEnrich}.}

\item{rlbaseRes}{A tibble containing results from RLBase samples.}

\item{yval}{Column to use for obtaining Y-axis values. Default: "stat_fisher_rl".
If NULL, no comparison with RLBase will be made.}
}
\value{
A named list of ggplot objects.
}
\description{
Plotting function for featureEnrich() results.
}
