% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment Test Results}
\usage{
plotEnrichment(
  object,
  modes = NULL,
  onlyCase = TRUE,
  onlyPOS = FALSE,
  splitby = c("none", "verdict", "condType"),
  limits = c(-10, 15),
  returnData = FALSE,
  ...
)
}
\arguments{
\item{object}{The tibble object obtained from running \code{featureEnrich}.}

\item{modes}{Which modes to include in plot? If empty, all modes included.}

\item{onlyCase}{If TRUE, only "case" predicted samples included.
Default: TRUE.}

\item{onlyPOS}{If TRUE, only "POS" labeled samples included. Default: FALSE.}

\item{splitby}{Metadata by which to split plots. Can be "none", "verdict",
or "condType".}

\item{limits}{Specify limits on data. Useful for controlling infinite
estimation of odds ratio
resulting from fisher's exact test. To remove limits, set c(-Inf, Inf).
Default: c(-10, 15).}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}

\item{...}{For internal use.}
}
\value{
A named list of ggplot objects.
}
\description{
Plot Enrichment Test Results
}
\examples{

# Example dataset with featureEnrich() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Plot enrichment
plotEnrichment(rlr)

# split by verdict
plotEnrichment(rlr, onlyCase = FALSE, splitby = "verdict")

# split by "condType" and only return plotting data
plotEnrichment(rlr, onlyCase = FALSE, splitby = "condType",
               returnData = TRUE)

# Only include DRIP-family
plotEnrichment(rlr,
    modes = c("DRIP", "DRIPc", "qDRIP", "sDRIP", "ssDRIP"),
    onlyCase = FALSE, splitby = "verdict"
)
}
