% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment Test Results}
\usage{
plotEnrichment(
  object,
  modes = NULL,
  onlyCase = TRUE,
  onlyPOS = FALSE,
  yval = "stat_fisher_rl"
)
}
\arguments{
\item{object}{The tibble obejct obtained from running \code{featureEnrich}.}

\item{modes}{Which modes to include in plot? If empty, all modes included.}

\item{onlyCase}{If TRUE, only "case" predicted samples included. Default: TRUE.}

\item{onlyPOS}{If TRUE, only "POS" labeled samples included. Default: FALSE.}

\item{yval}{Column to use for obtaining Y-axis values. Default: "stat_fisher_rl".
If NULL, no comparison with RLBase will be made.}
}
\value{
A named list of ggplot objects.
}
\description{
Plot Enrichment Test Results
}
