% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment Test Results}
\usage{
plotEnrichment(
  object,
  modes = NULL,
  pred_POS_only = TRUE,
  label_POS_only = FALSE,
  splitby = c("none", "prediction", "label"),
  limits = c(-10, 15),
  returnData = FALSE,
  ...
)
}
\arguments{
\item{object}{The tibble object obtained from running \code{featureEnrich}.}

\item{modes}{Which modes to include in plot? If empty, all modes included.}

\item{pred_POS_only}{If TRUE, only "POS" predicted samples included.
Default: TRUE.}

\item{label_POS_only}{If TRUE, only "POS" labeled samples included. Default: FALSE.}

\item{splitby}{Metadata by which to split plots. Can be "none", "prediction",
or "label".}

\item{limits}{Specify limits on data. Useful for controlling infinite
estimation of odds ratio
resulting from fisher's exact test. To remove limits, set c(-Inf, Inf).
Default: c(-10, 15).}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}

\item{...}{For internal use.}
}
\value{
A named list of ggplot objects.
}
\description{
Creates a list of plots, one for each annotation database which display the
user-supplied sample in comparison to the samples in RLBase. This
will only work if using non-custom annotations.
}
\examples{

# Example dataset with featureEnrich() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Plot enrichment
plotEnrichment(rlr)

# split by prediction
plotEnrichment(rlr, pred_POS_only = FALSE, splitby = "prediction")

# split by "label" and only return plotting data
plotEnrichment(rlr,
    pred_POS_only = FALSE, splitby = "label",
    returnData = TRUE
)

# Only include DRIP-family
plotEnrichment(rlr,
    modes = c("DRIP", "DRIPc", "qDRIP", "sDRIP", "ssDRIP"),
    pred_POS_only = FALSE, splitby = "prediction"
)
}
