% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLFSRes}
\alias{plotRLFSRes}
\title{Plot Perm Test results}
\usage{
plotRLFSRes(object, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run.}

\item{...}{Additional parameters passed to \code{ggplot}.}
}
\value{
A ggplot object.
}
\description{
Plot Perm Test results
}
\examples{

# Example RLRanges dataset with analyzeRLFS() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Plot RLFS res
plotRLFSRes(rlr)
}
