% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLFSRes}
\alias{plotRLFSRes}
\title{Plot Perm Test results}
\usage{
plotRLFSRes(object, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run.}

\item{...}{Additional parameters passed to \code{ggplot}.}
}
\value{
A ggplot object.
}
\description{
Plot Perm Test results
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Run RLSeq
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")
rlr <- RLSeq(rlr)

# Plot RLFS res
plotRLFSRes(rlr)
}
}
