% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLFSRes}
\alias{plotRLFSRes}
\title{Plot RLFS analysis results}
\usage{
plotRLFSRes(object, plotName = NULL, fft = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run. Alternatively,
can be the RLFS results object from an RLRanges.}

\item{plotName}{A Sample name used for plotting.
If blank, the RLRanges name is used.}

\item{fft}{If TRUE, the fourier transform of Z-score is plotted instead.
Default: FALSE.}

\item{...}{Additional parameters passed to \code{ggplot}.}
}
\value{
A ggplot object.
}
\description{
Plots the results of the R-loop-forming sequences analysis.
}
\examples{

# Example RLRanges dataset with analyzeRLFS() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Plot RLFS res
plotRLFSRes(rlr)

# Plot the Fourier transform instead
plotRLFSRes(rlr, fft=TRUE)

}
