% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLFSRes}
\alias{plotRLFSRes}
\title{Plot Perm Test results}
\usage{
plotRLFSRes(rlfsRes, ...)
}
\arguments{
\item{rlfsRes}{The list object generated from running \code{analyzeRLFS}.}

\item{...}{Additional parameters passed to \code{ggplot}.}
}
\value{
A ggplot object.
}
\description{
Simple function for plotting permTest
results using ggplot2.
}
