% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLRegionOverlap}
\alias{plotRLRegionOverlap}
\title{Plot RL-Region overlap with RLRanges}
\usage{
plotRLRegionOverlap(object, returnData = FALSE)
}
\arguments{
\item{object}{An RLRanges object with \code{rlRegionTest()} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}
}
\value{
A venn diagram ggplot object.
}
\description{
Plot RL-Region overlap with RLRanges
}
\examples{

# Example dataset with rlRegionTest() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Plot RL-Region overlap
plotRLRegionOverlap(rlr)

# Return data only
plotRLRegionOverlap(rlr, returnData = TRUE)
}
