% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLRegionOverlap}
\alias{plotRLRegionOverlap}
\title{Plot RL-Region overlap with RLRanges}
\usage{
plotRLRegionOverlap(object, returnData = FALSE, rlregions_table = NULL, ...)
}
\arguments{
\item{object}{An RLRanges object with \link{rlRegionTest} already run.}

\item{returnData}{If TRUE, plot data is returned instead of plotting.
Default: FALSE}

\item{rlregions_table}{The table of RLRegions to overlap sample ranges with.
Obtained from RLHUb using \link[RLHub:rlregions]{RLHub::rlregions_meta}. Loaded from RLHub if
not supplied. Default: NULL.}

\item{...}{Additional arguments passed to \link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram}}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object containing the venn diagram. Built
using \link[ggplotify:as.ggplot]{ggplotify::as.ggplot}.
}
\description{
Convenience function for plotting the overlap between RLRanges and R-loop
regions (RL regions) as calculated by \link{rlRegionTest}.
}
\examples{

# Example dataset with rlRegionTest() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Plot RL-Region overlap
plotRLRegionOverlap(rlr)

# Return data only
plotRLRegionOverlap(rlr, returnData = TRUE)
}
