% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLRegionOverlap}
\alias{plotRLRegionOverlap}
\title{Plot RL-Region overlap with RLRanges}
\usage{
plotRLRegionOverlap(object)
}
\arguments{
\item{object}{An RLRanges object with \code{rlRegionTest()} already run.}
}
\value{
A venn diagram ggplot object.
}
\description{
Plot RL-Region overlap with RLRanges
}
