% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCondition.R
\name{predictCondition}
\alias{predictCondition}
\title{Predict Condition}
\usage{
predictCondition(object, ...)
}
\arguments{
\item{object}{An RLRanges with \code{analyzeRLFS()} already run.}

\item{...}{Internal use only.}
}
\value{
A list containing the results of the fourier analysis and the model prediction.
}
\description{
Uses the results of analyzeRLFS() to predict whether a sample is "Case"
(Normal) or "Control" (RNaseH1-like). This is a useful quality metric when
interpreting R-loop mapping results.
}
\examples{

pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
rlr <- RLRanges(pks, genome="hg38", mode="DRIP")
rlr <- analyzeRLFS(rlr)
rlr <- predictCondition(rlr)

}
