% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCondition.R
\name{predictCondition}
\alias{predictCondition}
\title{Predict Condition}
\usage{
predictCondition(object, rlfsRes = NULL, ...)
}
\arguments{
\item{object}{An RLRanges object with \code{analyzeRLFS()} already run.
Ignored if rlfsRes provided.}

\item{rlfsRes}{If object not supplied, provide the rlfsRes list which is
obtained from \code{rlresult(object, "rlfsRes")}.}

\item{...}{Internal use only.}
}
\value{
An RLRanges object with predictions included.
}
\description{
Uses the results of analyzeRLFS() to predict whether a sample is "POS"
(robust R-loop mapping) or "NEG" (poor R-loop mapping).
}
\examples{

# Example data with analyzeRLFS already run
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# predict condition
rlr <- predictCondition(rlr)

# With rlfsRes
predRes <- predictCondition(rlfsRes = rlresult(rlr, "rlfsRes"))
}
