% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCondition.R
\name{predictCondition}
\alias{predictCondition}
\title{Predict Condition}
\usage{
predictCondition(object, ...)
}
\arguments{
\item{object}{An RLRanges object with \code{analyzeRLFS()} already run.}

\item{...}{Internal use only.}
}
\value{
An RLRanges object with predictions included.
}
\description{
Uses the results of analyzeRLFS() to predict whether a sample is "Case"
(robust R-loop mapping) or "Control" (poor R-loop mapping).
}
\examples{

# Example data with analyzeRLFS already run
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# predict condition
rlr <- predictCondition(rlr)

}
