% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCondition.R
\name{predictCondition}
\alias{predictCondition}
\title{Predict Condition}
\usage{
predictCondition(object, ...)
}
\arguments{
\item{object}{An RLRanges object with \code{analyzeRLFS()} already run.}

\item{...}{Internal use only.}
}
\value{
An RLRanges object with predictions included.
}
\description{
Uses the results of analyzeRLFS() to predict whether a sample is "Case"
(robust R-loop mapping) or "Control" (poor R-loop mapping).
}
\examples{
\dontrun{

# Example dataset
rlbase <- "https://rlbase-data.s3.amazonaws.com"
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
cvg <- file.path(rlbase, "coverage", "SRX1025890_hg38.bw")

# Get RLRanges object
rlr <- RLRanges(pks, coverage = cvg, genome = "hg38", mode = "DRIP")

# Get RLFS results
rlr <- analyzeRLFS(rlr)

# predict condition
rlr <- predictCondition(rlr)

}
}
