% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCondition.R
\name{predictCondition}
\alias{predictCondition}
\title{Predict Condition}
\usage{
predictCondition(rlfsRes, ...)
}
\arguments{
\item{rlfsRes}{The results list from running analyzeRLFS().}

\item{...}{Internal use only.}
}
\value{
A list containing the results of the fourier analysis and the model prediction.
}
\description{
Uses the results of analyzeRLFS() to predict whether a sample is "Case"
(Normal) or "Control" (RNaseH1-like). This is a useful quality metric when
interpreting R-loop mapping results.
}
\examples{

rlfsRes <- RLSeq::analyzeRLFS(RLSeq::SRX1025890_peaks, genome="hg38")
RLSeq::predictCondition(rlfsRes)

}
