% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate an RLSeq Report}
\usage{
report(object, reportPath = "rlreport.html", quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object.}

\item{reportPath}{A path indicating the report output HTML file.
Default: "rlreport.html"}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{rmarkdown::render()}}
}
\value{
TRUE
}
\description{
Generate an RLSeq Report
}
\examples{

# Example data with RLSeq() already run.
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Get a TMP file (only for example usae)
tmp <- tempfile()

# Generate the report
report(rlr, reportPath = tmp)
}
