% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate report}
\usage{
report(data, outputFile = "report.html", ...)
}
\arguments{
\item{data}{A named list containing \code{annotated_peaks}, \code{feature_overlaps},
and \code{rlfs_results}}

\item{outputFile}{A path indicating the report output HTML file.}

\item{...}{Arguments passed to \code{rmarkdown::render()}}
}
\description{
Generates RLSeq Quality Report
}
\examples{

RLSeq::makeReport(RLSeq::SRX1025890, outputFile = "report.html")

}
