% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{RLSeq Report}
\usage{
report(
  object,
  reportPath = "rlreport.html",
  intermediates_dir = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{object}{An RLRanges object.}

\item{reportPath}{A path indicating the report output HTML file.
Default: "rlreport.html"}

\item{intermediates_dir}{A directory indicating where intermediate
files should be stored during report building. If not set,
a random directory in \verb{tmp/} will be used. Default: NULL.}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to \code{rmarkdown::render()}}
}
\value{
TRUE
}
\description{
Builds an HTML report to showcase the results available in the supplied
\link{RLRanges} object (see also \link{RLResults}).
}
\examples{

# Example data with RLSeq() already run.
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Get a TMP file (only for example usae)
tmp <- tempfile()

# Generate the report
report(rlr, reportPath = tmp)
}
