% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRangesFromRLBase.R
\name{RLRangesFromRLBase}
\alias{RLRangesFromRLBase}
\title{Access RLBase samples as RLRanges}
\usage{
RLRangesFromRLBase(acc, rlsamples = NULL)
}
\arguments{
\item{acc}{The sample ID of the RLBase object. See the \code{rlsample} column
in \link[RLHub:rlbase_samples]{RLHub::rlbase_samples}.}

\item{rlsamples}{The tibble provided by \link[RLHub:rlbase_samples]{RLHub::rlbase_samples}.
Providing these data ahead of time speeds up this operation. Default: NULL.}
}
\value{
An RLRanges object with all results available.
}
\description{
Accessor function which returns any sample in
\href{https://gccri.bishop-lab.uthscsa.edu/rlbase/}{RLBase} as an
\link{RLRanges} object for use with RLSeq. For a full list of available
samples, see \link[RLHub:rlbase_samples]{RLHub::rlbase_samples}.
}
\examples{

rlr <- RLRangesFromRLBase("SRX1070676")

}
