% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRangesFromRLBase.R
\name{RLRangesFromRLBase}
\alias{RLRangesFromRLBase}
\title{Access RLBase samples as RLRanges}
\usage{
RLRangesFromRLBase(acc, rlsamples = NULL)
}
\arguments{
\item{acc}{The sample ID of the RLBase object. \code{rlsample} column in \code{RLHub::rlbase_samples()}.}

\item{rlsamples}{The tibble provided by \code{RLHub::rlbase_samples()}.
Providing these data ahead of time speeds up this operation. Default: NULL.}
}
\value{
An RLRanges object with all results available.
}
\description{
Accessor function which returns any sample in RLBase as an RLRanges object for
use with RLSeq.
}
\examples{

rlr <- RLRangesFromRLBase("SRX1070676")
}
