% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlRegionTest.R
\name{rlRegionTest}
\alias{rlRegionTest}
\title{R-Loop region test}
\usage{
rlRegionTest(peaks, genom = "hg38")
}
\arguments{
\item{peaks}{A GRanges object containing R-loop peaks}

\item{genome}{UCSC genome identifier to use. Can be only "hg38" currently.
If you have "hg19" peaks, please use RSeqR::liftUtil() to convert them.}
}
\value{
A named list containing the results of testing.
}
\description{
Tests the overlap of peaks within R-loop regions
}
\examples{

RSeqR::rlRegionTest(RSeqR::SRX1025890_peaks, genome="hg38")

}
