% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlRegionTest.R
\name{rlRegionTest}
\alias{rlRegionTest}
\title{R-Loop region test}
\usage{
rlRegionTest(object)
}
\arguments{
\item{object}{An RLRanges object with genome "hg38".}
}
\value{
An RLRanges object with test results included.
}
\description{
Tests the overlap of user-supplied ranges with R-loop regions (R-loop consensus
sites derived from meta-analysis of RLBase)
}
\examples{

# Example RLRanges data
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# RL Region Test
rlRegionTest(rlr)
}
