% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlRegionTest.R
\name{rlRegionTest}
\alias{rlRegionTest}
\title{R-Loop region test}
\usage{
rlRegionTest(object)
}
\arguments{
\item{object}{An RLRanges object with genome "hg38".}
}
\value{
An RLRanges object with test results included.
}
\description{
Tests the overlap of peaks within R-loop regions
}
\examples{
pks <- file.path(rlbase, "peaks", "SRX1025890_hg38.broadPeak")
rlr <- RLRanges(pks, genome="hg38", mode="DRIP", quiet=TRUE)

rlr <- rlRegionTest(rlr)
}
