% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLSeq.R
\name{RLSeq}
\alias{RLSeq}
\title{Run RLSeq}
\usage{
RLSeq(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{An RLRanges object.}

\item{quiet}{If TRUE, messages are suppressed. Default: FALSE.}

\item{...}{Arguments passed to analyzeRLFS.}
}
\value{
An RLRanges object with all results available.
}
\description{
Runs the full RLSeq pipeline.
}
\examples{

# Example RLRanges
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# Run RLSeq
rlr <- RLSeq(rlr)

}
