% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLSeq.R
\name{RLSeq}
\alias{RLSeq}
\title{RLSeq}
\usage{
RLSeq(object, quiet = FALSE, skip = NULL, ...)
}
\arguments{
\item{object}{An \link{RLRanges} object.}

\item{quiet}{If \code{TRUE}, messages are suppressed. Default: \code{FALSE}.}

\item{skip}{Analysis steps to skip.
Default: \code{NULL}. See \emph{details} for options.}

\item{...}{Arguments passed to \link{analyzeRLFS}.}
}
\value{
An \link{RLRanges} object with results available (see \link{rlresult}).
}
\description{
Executes the RLSeq analysis workflow.
}
\details{
The \code{RLSeq()} function does all of the following by default:
\enumerate{
\item \strong{RLFS Perm Test}. Runs the \link{analyzeRLFS} function to test the
enrichment of user-supplied ranges within R-loop-forming sequences.
\emph{Cannot be skipped.}
\item \strong{Predict Condition}. Runs the \link{predictCondition} function to
predict whether the user-supplied sample robustly maps R-loops or not.
\emph{Cannot be skipped.}
\item \strong{Noise analysis}. Runs the \link{noiseAnalyze} function to analyze the
signal-noise distribution within the supplied dataset. Skip with
\code{skip="noiseAnalyze"}.
\item \strong{Feature enrichment test}. Runs the \link{featureEnrich} function to
test the enrichment of user-supplied ranges within R-loop-relevant
genomic features. Skip with \code{skip="featureEnrich"}.
\item \strong{Transcript Feature Overlap}. Runs the \link{txFeatureOverlap} function to
get the overlap of transcript features and user-supplied peaks.
\item \strong{Correlation Analysis}. Runs the \link{corrAnalyze} function to test
the correlation of user-supplied R-loop signal with other samples in
RLBase around "gold-standard" R-loop regions.
Skip with \code{skip="corrAnalyze"}.
\item \strong{Gene annotation}. Runs the \link{geneAnnotation} function to find overlap
of genes with the user-supplied ranges. Skip with \code{skip="geneAnnotation"}.
\item \strong{R-loop Region Analysis}. Runs the \link{rlRegionTest} function to find
the overlap of user-supplied ranges with consensus R-loop sites
(RL-Regions). Skip with \code{skip="rlRegionTest"}.
}
}
\examples{

# Example RLRanges
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Run RLSeq
# `useMask=FALSE`, `ntime=10`, and `skip=` for demonstration purposes here.
rlr <- RLSeq(
    rlr,
    useMask = FALSE, ntimes = 10,
    skip = c(
        "featureEnrich", "corrAnalyze", "geneAnnotation", "rlRegionTest"
    )
)

}
