% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRX1025890}
\alias{SRX1025890}
\title{A named list that is generated by running RSeqR() containing all results from
the pipeline for the sample SRX1025890.}
\format{
An object of class \code{list} of length 6.
}
\usage{
SRX1025890
}
\description{
A named list that is generated by running RSeqR() containing all results from
the pipeline for the sample SRX1025890.
}
\keyword{datasets}
