% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRX1025890_peaks}
\alias{SRX1025890_peaks}
\title{GRanges. Peaks of running RSeq on SRX1025890}
\format{
An object of class \code{GRanges} of length 2902.
}
\usage{
SRX1025890_peaks
}
\description{
GRanges. Peaks of running RSeq on SRX1025890
}
\keyword{datasets}
