% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urlExists}
\alias{urlExists}
\title{Check if URL exists}
\usage{
urlExists(url)
}
\arguments{
\item{url}{URL to check}
}
\value{
logical. TRUE if status code 200, FALSE if not
}
\description{
Check if URL exists
}
